requirements.txt
fastapi==0.115.12
uvicorn==0.26.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
google-generativeai==0.7.0
python-multipart==0.0.9
python-dateutil==2.8.2
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.0.0
